+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/Model.kt
@@ -2,10 +2,12 @@
 package com.sourcegraph.cody.protocol_generated;
 
 data class Model(
-  val default: Boolean,
-  val codyProOnly: Boolean,
+  val model: String,
+  val usage: List<ModelUsage>,
+  val contextWindow: ModelContextWindow,
+  val clientSideConfig: ClientSideConfigParams? = null,
   val provider: String,
   val title: String,
-  val deprecated: Boolean,
+  val tags: List<ModelTag>,
 )