+++ vscode/webviews/chat/fixtures.ts
@@ -144,6 +144,7 @@ export const FIXTURE_TRANSCRIPT: Record<
 export const FIXTURE_USER_ACCOUNT_INFO: UserAccountInfo = {
     isCodyProUser: true,
     isDotComUser: true,
+    isOldStyleEnterpriseUser: false,
     user: {
         username: 'sqs',
         displayName: 'Quinn Slack',